+++ client/Packages/com.beamable/Common/Runtime/Content/SerializableDictionary.cs
@@ -1,4 +1,5 @@
 using Beamable.Content;
+using Beamable.Serialization;
 using System;
 using System.Collections;
 using System.Collections.Generic;
@@ -6,25 +7,48 @@ using UnityEngine;
 
 namespace Beamable.Common.Content
 {
-	[Serializable]
-	public class SerializableDictionaryStringToInt : SerializableDictionaryStringToSomething<int> { }
+	[Serializable] public class MapOfLong : SerializableDictionaryStringToLong { }
+	[Serializable] public class OptionalMapOfLong : Optional<MapOfLong> { }
+	[Serializable] public class MapOfInt : SerializableDictionaryStringToInt { }
+	[Serializable] public class OptionalMapOfInt : Optional<MapOfInt> { }
+	[Serializable] public class MapOfString : SerializableDictionaryStringToString { }
+	[Serializable] public class OptionalMapOfString : Optional<MapOfString> { }
+
+	[Serializable] public class MapOfObject : SerializableDictionaryStringToObject { }
+	[Serializable] public class OptionalMapOfObject : Optional<MapOfObject> { }
+	[Serializable] public class MapOfBool : SerializableDictionaryStringToBool { }
+	[Serializable] public class OptionalMapOfBool : Optional<MapOfBool> { }
+	[Serializable] public class MapOfShort : SerializableDictionaryStringToShort { }
+	[Serializable] public class OptionalMapOfShort : Optional<MapOfShort> { }
+	[Serializable] public class MapOfDouble : SerializableDictionaryStringToDouble { }
+	[Serializable] public class OptionalMapOfDouble : Optional<MapOfDouble> { }
+	[Serializable] public class MapOfFloat : SerializableDictionaryStringToFloat { }
+	[Serializable] public class OptionalMapOfFloat : Optional<MapOfFloat> { }
+	[Serializable] public class MapOfByte : SerializableDictionaryStringToByte { }
+	[Serializable] public class OptionalMapOfByte : Optional<MapOfByte> { }
+	[Serializable] public class MapOfGuid : SerializableDictionaryStringToGuid { }
+	[Serializable] public class OptionalMapOfGuid : Optional<MapOfGuid> { }
+	[Serializable] public class SerializableDictionaryStringToInt : SerializableDictionaryStringToSomething<int> { }
+	[Serializable] public class SerializableDictionaryStringToBool : SerializableDictionaryStringToSomething<bool> { }
+	[Serializable] public class SerializableDictionaryStringToShort : SerializableDictionaryStringToSomething<short> { }
+	[Serializable] public class SerializableDictionaryStringToDouble : SerializableDictionaryStringToSomething<double> { }
+	[Serializable] public class SerializableDictionaryStringToFloat : SerializableDictionaryStringToSomething<float> { }
+	[Serializable] public class SerializableDictionaryStringToByte : SerializableDictionaryStringToSomething<byte> { }
+	[Serializable] public class SerializableDictionaryStringToGuid : SerializableDictionaryStringToSomething<Guid> { }
 
 	[Serializable]
 	public class SerializableDictionaryStringToLong : SerializableDictionaryStringToSomething<long>
 	{
 		public SerializableDictionaryStringToLong() { }
 
-		public SerializableDictionaryStringToLong(IDictionary<string, long> existing)
-		{
-			foreach (var kvp in existing)
-			{
-				Add(kvp.Key, kvp.Value);
-			}
-		}
+		public SerializableDictionaryStringToLong(IDictionary<string, long> existing) : base(existing) { }
 	}
 
 	[Serializable]
 	public class OptionalSerializableDictionaryStringToString : Optional<SerializableDictionaryStringToString> { }
+
+	[Serializable]
+	public class OptionalSerializableDictionaryStringToSomething<T> : Optional<SerializableDictionaryStringToSomething<T>> { }
 
 	[Serializable]
 	public class SerializableDictionaryStringToString : SerializableDictionaryStringToSomething<string>
@@ -41,6 +65,12 @@ namespace Beamable.Common.Content
 		}
 	}
 
+	[Serializable]
+	public class SerializableDictionaryStringToObject : SerializableDictionaryStringToSomething<object>
+	{
+
+	}
+
 	public static class SerializableDictionaryStringToStringExtensions
 	{
 		public static SerializableDictionaryStringToString ToSerializable(this IDictionary<string, string> data) =>
@@ -51,6 +81,19 @@ namespace Beamable.Common.Content
 	public class SerializableDictionaryStringToSomething<T> : SerializableDictionary<string, T>, IDictionaryWithValue
 	{
 		public Type ValueType => typeof(T);
+
+		public SerializableDictionaryStringToSomething()
+		{
+
+		}
+
+		public SerializableDictionaryStringToSomething(IDictionary<string, T> existing)
+		{
+			foreach (var kvp in existing)
+			{
+				Add(kvp.Key, kvp.Value);
+			}
+		}
 	}
 
 	public interface IDictionaryWithValue : IDictionary, IIgnoreSerializationCallbacks