+++ client/Packages/com.beamable/CHANGELOG.md
@@ -11,7 +11,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Editor tooling for `SerializableDictionaryStringToSomething<T>` has context menu
+
+
+### [1.3.2]
+
+### Added
+- `PaymentsService` now supports receipt verification through the `VerifyReceipt` method.
+- Logs improvements in Requesters.
+
+
+## [1.3.1]
+### Added
 - A leaderboard can now be frozen using `LeaderboardService.FreezeLeaderboard` method to prevent additional scores to be submitted.
+
+### Changed
+- Editor tooling for `SerializableDictionaryStringToSomething<T>` now supports all subtypes
 
 ### Fixed
 - iOS builds will no longer overwrite the Beamable user language preference.